Fixes bug 398536...
Adds an option "MarkerMethod"

Since this is the OpenCascadeLink we want the default to use the open cascade boundary marker computation in general. For the STL file import that is problematic as each surface element will get it's own marker. I have added an option "MarkerMethod" with the default using OpenCascade, "ElementMesh" will use ToBoundaryMesh to compute the markers and None does not include any markers. This is documented.

Needs["NDSolve`FEM`"]
Needs["OpenCascadeLink`"]
shape = OpenCascadeShapeImport[
   FileNameJoin[{$InstallationDirectory, "SystemFiles", "Components",
     "PDEModels", "SupportFiles", "ClassicalMuffler.stl"}]];
bmesh = OpenCascadeShapeSurfaceMeshToBoundaryMesh[shape,
  "MarkerMethod" -> "ElementMesh"]